Extract BT readers
